+++ crates/project/src/project.rs
@@ -8,9 +8,10 @@ use moon_config::package::PackageJson;
 use moon_config::tsconfig::TsConfigJson;
 use moon_config::{
     format_figment_errors, FilePath, GlobalProjectConfig, ProjectConfig, ProjectID, TaskID,
+    TypeScriptConfig,
 };
 use moon_logger::{color, debug, trace, Logable};
-use moon_utils::path;
+use moon_utils::{fs, path, string_vec};
 use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
 use std::path::{Path, PathBuf};
@@ -339,6 +340,38 @@ impl Project {
         })
     }
 
+    // Automatically create missing config files when we are syncing project references.
+    #[track_caller]
+    pub async fn create_tsconfig_json(
+        &self,
+        typescript_config: &TypeScriptConfig,
+        workspace_root: &Path,
+    ) -> Result<(), ProjectError> {
+        let tsconfig_path = self.root.join(&typescript_config.project_config_file_name);
+
+        if !self.tsconfig_json.initialized() && !tsconfig_path.exists() {
+            let tsconfig_options_path =
+                workspace_root.join(&typescript_config.root_options_config_file_name);
+
+            let json = TsConfigJson {
+                extends: Some(path::to_virtual_string(
+                    &path::relative_from(&tsconfig_options_path, &self.root).unwrap(),
+                )?),
+                include: Some(string_vec!["**/*"]),
+                references: Some(vec![]),
+                ..TsConfigJson::default()
+            };
+
+            fs::write_json(&tsconfig_path, &json, true).await?;
+
+            self.tsconfig_json
+                .set(json)
+                .expect("Failed to create tsconfig.json");
+        }
+
+        Ok(())
+    }
+
     /// Return a list of project IDs this project depends on.
     pub fn get_dependencies(&self) -> Vec<ProjectID> {
         let mut depends_on = vec![];
@@ -410,18 +443,21 @@ impl Project {
 
     /// Load and parse the package's `tsconfig.json` if it exists.
     #[track_caller]
-    pub async fn load_tsconfig_json(&self, tsconfig_name: &str) -> Result<bool, ProjectError> {
+    pub async fn load_tsconfig_json(
+        &self,
+        typescript_config: &TypeScriptConfig,
+    ) -> Result<bool, ProjectError> {
         if self.tsconfig_json.initialized() {
             return Ok(true);
         }
 
-        let tsconfig_path = self.root.join(tsconfig_name);
+        let tsconfig_path = self.root.join(&typescript_config.project_config_file_name);
 
         if tsconfig_path.exists() {
             trace!(
                 target: self.get_log_target(),
                 "Loading {} in {}",
-                color::file(tsconfig_name),
+                color::file(&typescript_config.project_config_file_name),
                 color::path(&self.root),
             );